MAIL/pyproject.toml
[project]
name = "mail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai==0.28.0",
    "python-dotenv>=1.1.1",
]
